+++ CHANGELOG.md
@@ -10,8 +10,11 @@ Breaking changes:
 * None.
 
 Other notable changes:
+* Added support for incoming cookies.
 * Notice promptly when clients drop connections, instead of letting them time
   out (and end up getting a write-after-close error).
+* Cleaned up the request logging code, including details of what lands in the
+  system log.
 
 ### v0.6.2 -- 2024-01-22